tests/requirements.txt
# python:
coveralls
# disable this if 'pkg_resources.VersionConflict' error happens:
flake8 < 3.5.0
nose
# stick the version to avoild the error:
# module 'pycodestyle' has no attribute 'break_around_binary_operator, etc.
pycodestyle<2.4.0
pylint

# ansible:
ansible-lint
# .. todo:: podman, https://github.com/containers/python-podman
#docker
#podman
molecule
yamllint
